requirements.txt
Django==3.1
djangorestframework==3.11.1
django-rest-auth==0.9.5
django-rest-knox==4.1.0
django-vote==2.2.0
django-rest-registration==0.5.6
djangorestframework-simplejwt==4.4.0
